Rename referenceObject to Ref
