+++ src/domain/repositories/DataElementRepository.ts
@@ -1,5 +1,5 @@
 import DataElement from "../entities/DataElement";
-import { Id } from "../entities/ReferenceObject";
+import { Id } from "../entities/Ref";
 
 export default interface DataElementRepository {
     getByDataSet(dataSetId: Id): Promise<DataElement[]>
