+++ src/domain/repositories/OrgUnitRepository.ts
@@ -1,4 +1,4 @@
-import { Id } from "../entities/ReferenceObject";
+import { Id } from "../entities/Ref";
 import { OrgUnit } from "../entities/OrgUnit";
 
 export default interface OrgUnitRepository {
