+++ src/domain/usecases/GetDataElementsUseCase.ts
@@ -1,6 +1,6 @@
 import DataElement from "../entities/DataElement";
 import DataElementRepository from "../repositories/DataElementRepository";
-import { Id } from "../entities/ReferenceObject";
+import { Id } from "../entities/Ref";
 
 export class GetDataElementsUseCase implements GetDataElementsUseCase {
     constructor(private dataElementRepository: DataElementRepository) { }
